+++ lib/linkhut_web/controllers/settings/stats_html.ex
@@ -1,0 +1,7 @@
+defmodule LinkhutWeb.Settings.StatsHTML do
+  use LinkhutWeb, :html
+
+  import LinkhutWeb.SettingsComponents
+
+  embed_templates "stats_html/*"
+end